Add logout component
Created logout component that is rendered inside the dashboards navbar.
